Refactor legend to incorporate gradient legend
